Expand ETD integration test
Connects to sul-dlss/infrastructure-integration-test#43

This in-progress branch, a separate-but-related implementation to that in the `moar-etd-steps` branch, opens up the ability to test ETDs all the way through application-based submission, the submit workflow, the accessioning workflow, and the release workflow. It verifies that the ETD looks as expected in Argo and PURL. It is blocked from going forward due to https://github.com/sul-dlss/infrastructure-integration-test/issues/43#issuecomment-671439951 and is not a high priority for the current work cycle.