Full host names
Ensure that Settings.preassembly_host includes stanford.edu so that scp
command works reliably with the ssh config in the README.md:

```
Host *.stanford.edu
  GSSAPIDelegateCredentials yes
```

This will only delegate for the full hostname. I also fixed a typo and
provided a little bit more of a hint of where to find ETD
username/password.