Press Enter key to submit search term
There are multiple buttons on the page that contain the text "Search" (the other is the "Released to Searchworks" facet).  But the current locator for finding and filling in the search box still finds exactly one element, and pressing the Enter key in it after filling it in submits the search form successfully.

(Finding based on exact text would also be tricky: I noticed when debugging that the actual text of the search box submission button, from the perspective of Capybara::Node::Element#text is "Search search icon" because of alt text and icon, even when the text is limited to what Capybara considers visible.)