Refactor reload page helper
The `#reload_page_until_timeout!` method has become a bit of a beast whose behavior depends primarily on which arguments it is provided and the order the code checks those args. It is clear now that the best way to extend this method for all the different use cases it supports is to give it a single default behavior (check text on a page) and to allow consumers to provide a block for special behavior. This is a clearer way to extend Ruby methods to new use cases and is more Rubyish.