Prefer setting proxyjump in config over command-line options
Why? Doing so allows us to use a more sophisticated setup using SSH multiplexing, holding a connection open for a short period of time to reduce the overall number of MFA pushes. This change depends on all of us devs having some common SSH config which will be shared and documented elsewhere.