Adding additional sleeps
While testing an [optimization] I made in moab-versioning I noticed that
the h2_versioned_object_creation feature spec can fail in several
different ways, some of which cause Honeybadger alerts that halt the spec, others
that cause an Honeybadger alert that doesn't cause the spec to halt, and
others that are just spec failures related to text not showing up on the
screen and throw no Honeybadger alert.

I increased the existing sleep, and added some additional sleeps between
when versions were created in the hopes that it would help. It did seem
to help a bit, but the tests can still fail occasionally.

[optimization]: https://github.com/sul-dlss/moab-versioning/pull/384